requirements.txt
dash
pandas
numpy
plotly
xgboost
scikit-learn
openpyxl
azure-storage-blob
